refactor: divide role
